+++ rust/host/k16-vm/src/computer/devices.rs
@@ -1,6 +1,7 @@
 mod bios;
 mod control;
 mod framebuffer;
+mod keyboard;
 mod serial;
 mod storage;
 mod text_display;
@@ -9,6 +10,9 @@ mod timer;
 pub(crate) use bios::BiosFlashDevice;
 pub(crate) use control::ComputerControlDevice;
 pub(crate) use framebuffer::FramebufferDevice;
+pub(crate) use keyboard::validate_event as validate_keyboard_event;
+pub(crate) use keyboard::KeyboardDevice;
+pub use keyboard::KeyboardEvent;
 pub(crate) use serial::{DebugSerialDevice, SerialInputDevice};
 pub(crate) use storage::{
     K16VolumeFileStorageMedia, StoragePortControllerSnapshot, StoragePortDevice,