+++ rust/host/k16-vm/src/computer/machine/construction.rs
@@ -1,7 +1,7 @@
 use super::ComputerMachine;
 use crate::computer::devices::{
     ComputerControlDevice, DebugSerialDevice, FramebufferDevice, K16VolumeFileStorageMedia,
-    SerialInputDevice, StoragePortDevice, TextDisplayDevice, TimerDevice,
+    KeyboardDevice, SerialInputDevice, StoragePortDevice, TextDisplayDevice, TimerDevice,
 };
 use crate::computer::profile::{
     validate_profile_v2, ComputerHardwareConfig, ComputerHardwareDevice, ComputerMachineProfile,
@@ -31,6 +31,7 @@ pub(super) fn from_profile(
         framebuffer0_device_id: device_ids.framebuffer,
         storage0_device_id: device_ids.storage_port,
         timer0_device_id: device_ids.timer,
+        keyboard0_device_id: device_ids.keyboard,
         bios_flash_device_id: None,
         cpus: Vec::new(),
         boot_cpu: None,
@@ -46,6 +47,7 @@ struct ConstructionDeviceIds {
     framebuffer: Option<MmioDeviceId>,
     storage_port: Option<MmioDeviceId>,
     timer: Option<MmioDeviceId>,
+    keyboard: Option<MmioDeviceId>,
 }
 
 fn hardware_table_entries(profile: &ComputerMachineProfile) -> Vec<HardwareTableEntry> {
@@ -89,6 +91,7 @@ fn map_hardware_device(
         ComputerHardwareDevice::Framebuffer => Box::new(FramebufferDevice::new()),
         ComputerHardwareDevice::StoragePort(config) => Box::new(storage_port_device(config)?),
         ComputerHardwareDevice::Timer => Box::new(TimerDevice::new()),
+        ComputerHardwareDevice::Keyboard => Box::new(KeyboardDevice::new()),
     };
     bus.map_mmio(hardware.mmio_base, device)
 }
@@ -115,6 +118,7 @@ impl ConstructionDeviceIds {
             ComputerHardwareDevice::Framebuffer => self.framebuffer = Some(device_id),
             ComputerHardwareDevice::StoragePort(_) => self.storage_port = Some(device_id),
             ComputerHardwareDevice::Timer => self.timer = Some(device_id),
+            ComputerHardwareDevice::Keyboard => self.keyboard = Some(device_id),
         }
     }
 }